+++ CadEditor/Plugin.cs
@@ -48,9 +48,9 @@ namespace CadEditor
         Bitmap makeImageRectangle(byte[] videoChunk, byte[] pallete, int subPalIndex, float scale, bool scaleAccurate = true, bool withAlpha = false);
 
         Bitmap[] makeObjects(ObjRec[] objects, Bitmap[][] objStrips, float scale, MapViewType drawType, int constantSubpal = -1);
-        Bitmap[] makeObjects(byte videoPageId, byte tilesId, byte palId, float scale, MapViewType drawType, int constantSubpal = -1);
+        Bitmap[] makeObjects(int videoPageId, int tilesId, int palId, float scale, MapViewType drawType, int constantSubpal = -1);
         Bitmap makeObject(int index, ObjRec[] objects, Bitmap[][] objStrips, float scale, MapViewType drawType, int constantSubpal = -1);
-        Bitmap makeObjectsStrip(byte videoPageId, byte tilesId, byte palId, float scale, MapViewType drawType, int constantSubpal = -1);
+        Bitmap makeObjectsStrip(int videoPageId, int tilesId, int palId, float scale, MapViewType drawType, int constantSubpal = -1);
 
         Bitmap makeScreen(int scrNo, int levelNo, int videoNo, int bigBlockNo, int blockNo, int palleteNo, float scale = 2.0f, bool withBorders = true);
